test(cli): split diagnostics integration coverage
